Factor out date param handling
