+++ app/models/concerns/announcement/block/has_end_date.rb
@@ -1,0 +1,21 @@
+class Announcement
+  class Block
+    module HasEndDate
+      extend ActiveSupport::Concern
+
+      included do
+        before_create :end_date_param
+      end
+
+      def end_date_param
+        if self.parameters["end_date"].present?
+          DateTime.parse(self.parameters["end_date"])
+        else
+          self.parameters["end_date"] ||= Time.now.to_s
+        end
+      end
+    end
+
+  end
+
+end
