+++ app/models/concerns/announcement/block/has_flexible_start_date.rb
@@ -1,0 +1,13 @@
+class Announcement
+  class Block
+    module HasFlexibleStartDate
+      extend ActiveSupport::Concern
+
+      def start_date_param
+        self.parameters["start_date"].present? ? DateTime.parse(self.parameters["start_date"]) : nil
+      end
+    end
+
+  end
+
+end
